requirements.txt
streamlit
pillow
pypdf
streamlit-sortables
